+++ src/core/__tests__/config.command.spec.ts
@@ -1,19 +1,21 @@
 import { Test } from '@nestjs/testing'
+import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 import { ConfigCommand } from '~/core/config.command'
-import { config, configModule } from '~/core/config.module'
+import { config } from '~/core/config.module'
+
+import { CliModule } from '~/cli.module'
 
 describe(ConfigCommand.name, () => {
   let configCommand: ConfigCommand
 
   beforeEach(async () => {
     const module = await Test.createTestingModule({
-      imports: [configModule],
-      providers: [ConfigCommand],
+      imports: [CliModule],
     }).compile()
 
     configCommand = module.get(ConfigCommand)
-    jest.spyOn(configCommand['logger'], 'log').mockImplementation(jest.fn())
+    vi.spyOn(configCommand['logger'], 'log').mockImplementation(vi.fn())
   })
 
   it('should log current config', () => {